UI: Show the documentation on every page.
